requirements.txt
torch==2.5.1
typer==0.15.1
fastapi==0.115.6
uvicorn==0.34.0
scikit_learn==1.6.0
matplotlib==3.10.0
hydra-core==1.3.2
torchvision
python-dotenv
wandb
pytest
coverage
pre-commit
